+++ app/src/main/java/com/example/carolinereid/musicexchange/Guitar.java
@@ -10,7 +10,7 @@ import com.example.carolinereid.musicexchange.Behaviours.Sellable;
 public class Guitar extends Instrument {
     private int numStrings;
 
-    public Guitar(String material, String colour, String type, int costPrice, int retailPrice, int numStrings) {
+    public Guitar(String material, String colour, String type, int costPrice, double retailPrice, int numStrings) {
         super(material, colour, type, costPrice, retailPrice);
         this.numStrings = numStrings;
     }
